# CrawlerAPI/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crawler API
  version: 1.0.0
  description: |
    OpenAPI specification for the Crawler API. Generated from Spring controller mappings
    under `src/main/java/com/crawlerApi/api`. Many endpoints require an authenticated account
    context (JWT) and return domain-specific DTOs for domains, audits, pages, elements, and test records.
servers:
  - url: http://localhost:9080
    description: Local default server port from application.properties
security:
  - bearerAuth: []
...
paths:
...
  /v1/auditrecords/{audit_record_id}/report:
    post:
      tags: [Audit Records]
      summary: Request audit record report
      description: Request a report for the given audit record; creates/links account by email.
      parameters:
        - $ref: '#/components/parameters/audit_record_id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Account' }
      responses:
        '200': { description: Report requested }
        '401': { description: Authentication required }
        '404': { description: Audit record not found }
  /v1/auditrecords/{audit_record_id}/pages:
    get:
      tags: [Audit Records]
      summary: Get pages for audit record
      parameters:
        - $ref: '#/components/parameters/audit_record_id'
      responses:
        '200': { description: Pages returned }
  /v1/auditrecords/{audit_record_id}/elements:
    get:
      tags: [Audit Records]
      summary: Get elements for audit record
      parameters:
        - $ref: '#/components/parameters/audit_record_id'
      responses:
        '200': { description: Elements returned }
  /v1/auditrecords/{audit_record_id}/stats:
    get:
      tags: [Audit Records]
      summary: Get stats for audit record
      parameters:
        - $ref: '#/components/parameters/audit_record_id'
      responses:
        '200': { description: Stats returned }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    audit_record_id:
      in: path
      name: audit_record_id
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
...
    Account:
      type: object
      description: User account (Auth0-linked)
      properties:
        id: { type: integer, format: int64 }
        userId: { type: string, description: "Auth0 user ID" }
        email: { type: string, format: email }
        name: { type: string }
        apiToken: { type: string }
        onboardedSteps: { type: array, items: { type: string } }
      additionalProperties: true
...